splits text out to its own component
